requirements.txt
amqp==1.4.9
anyjson==0.3.3
Babel==2.3.4
billiard==3.3.0.23
celery==3.1.23
Django==1.10.1
djangorestframework==3.4.6
flower==0.9.1
kombu==3.0.35
pytz==2016.6.1
tornado==4.2
